[FIX] connecting time place module algorithm bug
