+++ ulink/models/subject.js
@@ -1,39 +1,5 @@
 const pool = require('../modules/pool');
-
-const connectTimePlace = (subjects) => {
-    const res = [];
-    let timePlace;
-    let temp = {
-        'subjectIdx': 'n'
-    };
-    for (let subject of subjects) {
-        if (temp.subjectIdx != subject.subjectIdx) {
-            if (temp.subjectIdx != 'n') {
-                res.push(Object.assign(temp, timePlace));
-            }
-            temp = {
-                'subjectIdx': subject.subjectIdx,
-                'subjectCode': subject.subjectCode,
-                'name': subject.name,
-                'professor': subject.professor,
-                'credit': subject.credit,
-                'course': subject.course
-            };
-            timePlace = {
-                'startTime': [],
-                'endTime': [],
-                'day': [],
-                'content': []
-            };
-        }
-        timePlace.startTime.push(subject.startTime);
-        timePlace.endTime.push(subject.endTime);
-        timePlace.day.push(subject.day);
-        timePlace.content.push(subject.content);
-    }
-
-    return res;
-}
+const connect = require('../modules/connect');
 
 const checkOnDay = (onDay, day) => {
     if (onDay === undefined) return true;
@@ -73,7 +39,7 @@ const subject = {
         ON q1.subjectIdx = q2.subjectIdx ORDER BY q1.name, q2.startTime`;
         try {
             const subjects = await pool.queryParam(query2);
-            return connectTimePlace(subjects);
+            return await connect.connectTimePlace(subjects);
         } catch (err) {
             if (err.errno == 1062) {
                 console.log('getSubject ERROR : ', err.errno, err.code);
@@ -116,7 +82,7 @@ const subject = {
         ON q1.subjectIdx = q2.subjectIdx ORDER BY q1.name, q2.startTime`;
         try {
             const subjects = await pool.queryParam(query2);
-            const subjectTimePlace = connectTimePlace(subjects);
+            const subjectTimePlace = await connect.connectTimePlace(subjects);
             if(condition2.onDay === undefined && condition2.offDay === undefined){
                 return subjectTimePlace;
             }
@@ -170,7 +136,7 @@ const subject = {
         ON q1.subjectIdx = q2.subjectIdx ORDER BY q1.name, q2.startTime`;
         try {
             const subjects = await pool.queryParam(query2);
-            return connectTimePlace(subjects);
+            return await connect.connectTimePlace(subjects);
         } catch (err) {
             if (err.errno == 1062) {
                 console.log('getCourseSubject ERROR : ', err.errno, err.code);
@@ -194,7 +160,7 @@ const subject = {
         ON q1.subjectIdx = q2.subjectIdx ORDER BY q1.name, q2.startTime`;
         try {
             const subjects = await pool.queryParam(query2);
-            return connectTimePlace(subjects);
+            return await connect.connectTimePlace(subjects);
         } catch (err) {
             if (err.errno == 1062) {
                 console.log('getGradeSubject ERROR : ', err.errno, err.code);
@@ -218,7 +184,7 @@ const subject = {
         ON q1.subjectIdx = q2.subjectIdx ORDER BY q1.name, q2.startTime`;
         try {
             const subjects = await pool.queryParam(query2);
-            return connectTimePlace(subjects);
+            return await connect.connectTimePlace(subjects);
         } catch (err) {
             if (err.errno == 1062) {
                 console.log('getSearchSubject ERROR : ', err.errno, err.code);
